terminado

+++ src/ar/edu/unahur/obj2/clubYate/YateMotor.java
@@ -9,7 +9,7 @@ public class YateMotor extends Yate {
     private Integer autonomia;
 
     public YateMotor(String nombre, String duenio, Double manga,
-                     Double calado, Integer eslora, Integer tripulacion, Integer peso, Integer desplazamiento,
+                     Double calado, Double eslora, Integer tripulacion, Integer peso, Integer desplazamiento,
                      String propulsion, Integer potencia, Integer velocidad, Integer autonomia) {
         this.nombre = nombre;
         this.duenio = duenio;
